fix: 워크스페이스 전체 clippy 에러 해결 (117개 → 0개)
## Clippy 수정 (117개 에러 100% 해결)

### trader-core (56개)
- type_complexity, manually_reimplementing_div_ceil 등

### trader-cli (10개)
- unnecessary_cast (chart_gen.rs: 3개)
- len_zero (migrate.rs: 1개)
- derivable_impls, needless_borrow (strategy_test.rs: 2개)
- wrong_self_convention (backtest.rs: 1개)
- cloned_ref_to_slice_refs, unnecessary_lazy_evaluations (main.rs: 2개)
- new_without_default (engine.rs: 1개)

### trader-execution (컴파일 에러 수정)
- ConversionConfig 구조체 초기화 패턴 수정

### 기타
- trader-analytics, trader-api, trader-collector, trader-strategy 등 다수
- doc_lazy_continuation, unused_variables 수정

## 설정 업데이트
- 에이전트 설정 (.claude/agents/) 업데이트
- 커스텀 에이전트 추가 (db-reviewer)
- 후크 스크립트 개선

+++ crates/trader-analytics/src/trigger_calculator.rs
@@ -414,7 +414,7 @@ mod tests {
 
         // 하락 후 반등 패턴 (RSI가 30 → 60으로 상승)
         for i in 0..10 {
-            let price = dec!(100) - Decimal::from(i as i64);
+            let price = dec!(100) - Decimal::from(i);
             klines.push(create_test_kline(
                 price,
                 price + dec!(1),
@@ -427,7 +427,7 @@ mod tests {
 
         // 반등 시작
         for i in 10..30 {
-            let price = dec!(90) + Decimal::from(i as i64 - 10) * dec!(2);
+            let price = dec!(90) + Decimal::from(i - 10) * dec!(2);
             klines.push(create_test_kline(
                 price,
                 price + dec!(2),
@@ -450,7 +450,7 @@ mod tests {
         let mut klines = Vec::new();
 
         for i in 0..30 {
-            let price = dec!(100) + Decimal::from(i as i64);
+            let price = dec!(100) + Decimal::from(i);
             klines.push(create_test_kline(
                 price,
                 price + dec!(2),
